refactor: fix duplicated editor package
