+++ app/src/main/java/com/github/kitakkun/noteapp/MainApplication.kt
@@ -3,7 +3,7 @@ package com.github.kitakkun.noteapp
 import android.app.Application
 import com.github.kitakkun.noteapp.data.dataModule
 import com.github.kitakkun.noteapp.di.appModule
-import com.github.kitakkun.noteapp.editor.editor.editorModule
+import com.github.kitakkun.noteapp.editor.editorModule
 import com.github.kitakkun.noteapp.finder.finderModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
