+++ app/src/main/java/com/github/kitakkun/noteapp/navigation/PageRouter.kt
@@ -12,7 +12,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.navigation
-import com.github.kitakkun.noteapp.editor.editor.EditorPage
+import com.github.kitakkun.noteapp.editor.EditorPage
 import com.github.kitakkun.noteapp.finder.FinderPage
 import com.github.kitakkun.noteapp.license.LicensePage
 import com.github.kitakkun.noteapp.setting.SettingPage
