+++ editor/src/main/java/com/github/kitakkun/noteapp/editor/ext/Color.kt
@@ -1,4 +1,4 @@
-package com.github.kitakkun.noteapp.editor.editor.ext
+package com.github.kitakkun.noteapp.editor.ext
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
